task

+++ app/Http/Controllers/TasksController.php
@@ -17,21 +17,22 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
 
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            $data += $this->counts($user);
+            return view('tasks.index', $data);
+        }else {
+            return view('welcome');
+        }
     }
     
-    public function welcome()
-    {
-        $tasks = Task::all();
-
-        return view('welcome', [
-            'tasks' => $tasks,
-        ]);
-    }
 
     /**
      * Show the form for creating a new resource.
@@ -54,12 +55,14 @@ class TasksController extends Controller
      */
     public function store(Request $request)
     {
+        $user = \Auth::user();
         $this->validate($request, [
         'status' => 'required|max:10',   // add
         'content' => 'required|max:191',
         ]);
         
         $task = new Task;
+        $task->user_id = $user->id;
         $task->status = $request->status;    // add
         $task->content = $request->content;
         $task->save();
@@ -75,11 +78,12 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-        $task = Task::find($id);
-
+       $task = Task::find($id);
+       if (\Auth::user()->id === $task->user_id) {
         return view('tasks.show', [
             'task' => $task,
         ]);
+       }
     }
 
     /**
@@ -91,10 +95,11 @@ class TasksController extends Controller
     public function edit($id)
     {
         $task = Task::find($id);
-
-        return view('tasks.edit', [
-            'task' => $task,
+         if (\Auth::user()->id === $task->user_id) {
+            return view('tasks.edit',[
+                'task' => $task,
         ]);
+        }
     }
 
     /**
@@ -127,9 +132,12 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
+       $task = \App\Task::find($id);
+
+        if (\Auth::user()->id === $task->user_id) {
+            $task->delete();
 
         return redirect('/');
+        }
     }
 }
